Fixes image callback issues and removes redundancy
Image callback wasn't functional since the bothook didn't call the drawmanager to set an image to call back. Also re-wrote the methods for checking if an item in the inventory is selected. This removes the odd global volatile and also removes creating a new image item every time we check a item in the inventory. Instead it now creates the image once and then loops through the components to see if any are highlighted.